Bug1189466: Move run_playback into MediaTestCase so that tests can be separated easily.
